requirements.txt
requests==2.18.2
SQLObject==3.3.0
lxml==3.5.0
telepot==12.1
unittest2==1.1.0
xmpppy==0.5.0rc1
